small refactor to cleanup code
