+++ music_assistant/controllers/webserver/controller.py
@@ -10,43 +10,22 @@ from __future__ import annotations
 import asyncio
 import hashlib
 import html
-import logging
 import os
 import urllib.parse
 from collections.abc import Awaitable, Callable
 from concurrent import futures
-from contextlib import suppress
 from functools import partial
 from typing import TYPE_CHECKING, Any, Final, cast
 from urllib.parse import parse_qs, quote, urlencode, urlparse, urlunparse
 
 import aiofiles
-from aiohttp import ClientTimeout, WSMsgType, web
+from aiohttp import ClientTimeout, web
 from mashumaro.exceptions import MissingField
 from music_assistant_frontend import where as locate_frontend
-from music_assistant_models.api import (
-    CommandMessage,
-    ErrorResultMessage,
-    MessageType,
-    SuccessResultMessage,
-)
-from music_assistant_models.auth import (
-    AuthProviderType,
-    AuthToken,
-    User,
-    UserAuthProvider,
-    UserRole,
-)
+from music_assistant_models.api import CommandMessage
+from music_assistant_models.auth import AuthProviderType, UserRole
 from music_assistant_models.config_entries import ConfigEntry, ConfigValueOption
 from music_assistant_models.enums import ConfigEntryType
-from music_assistant_models.errors import (
-    AuthenticationFailed,
-    AuthenticationRequired,
-    InsufficientPermissions,
-    InvalidCommand,
-    InvalidDataError,
-    InvalidToken,
-)
 
 from music_assistant.constants import (
     CONF_BIND_IP,
@@ -55,7 +34,7 @@ from music_assistant.constants import (
     RESOURCES_DIR,
     VERBOSE_LOG_LEVEL,
 )
-from music_assistant.helpers.api import APICommandHandler, api_command, parse_arguments
+from music_assistant.helpers.api import parse_arguments
 from music_assistant.helpers.audio import get_preview_stream
 from music_assistant.helpers.json import json_dumps, json_loads
 from music_assistant.helpers.util import get_ip_addresses
@@ -71,13 +50,11 @@ from .api_docs import (
 from .auth import AuthenticationManager
 from .helpers.auth_middleware import (
     get_authenticated_user,
-    get_current_token,
-    get_current_user,
     is_request_from_ingress,
-    set_current_token,
     set_current_user,
 )
 from .helpers.auth_providers import BuiltinLoginProvider
+from .websocket_client import WebsocketClientHandler
 
 if TYPE_CHECKING:
     from music_assistant_models.config_entries import ConfigValueType, CoreConfig
@@ -177,7 +154,7 @@ class WebserverController(CoreController):
             ),
         )
 
-    async def setup(self, config: CoreConfig) -> None:
+    async def setup(self, config: CoreConfig) -> None:  # noqa: PLR0915
         """Async initialize of module."""
         self.config = config
         # work out all routes
@@ -599,7 +576,7 @@ class WebserverController(CoreController):
             token_hash = hashlib.sha256(token.encode()).hexdigest()
             token_row = await self.auth.database.get_row("auth_tokens", {"token_hash": token_hash})
             if token_row:
-                await self.auth.revoke_token(token_row["token_id"], user)
+                await self.auth.database.delete("auth_tokens", {"token_id": token_row["token_id"]})
 
         return web.json_response({"success": True})
 
@@ -901,388 +878,6 @@ class WebserverController(CoreController):
                 {"success": False, "error": f"Setup failed: {e!s}"}, status=500
             )
 
-    # User Management API Methods
-
-    @api_command("auth/users", required_role="admin")
-    async def get_users(self) -> list[User]:
-        """
-        Get all users (admin only).
-
-        :return: List of user objects.
-        """
-        return await self.auth.list_users()
-
-    @api_command("auth/user", required_role="admin")
-    async def get_user(self, user_id: str) -> User | None:
-        """
-        Get user by ID (admin only).
-
-        :param user_id: The user ID.
-        :return: User object or None if not found.
-        """
-        return await self.auth.get_user(user_id)
-
-    @api_command("auth/user/create", required_role="admin")
-    async def create_user(
-        self,
-        username: str,
-        password: str,
-        role: str = "user",
-        display_name: str | None = None,
-        avatar_url: str | None = None,
-    ) -> User:
-        """
-        Create a new user with built-in authentication (admin only).
-
-        :param username: The username (minimum 3 characters).
-        :param password: The password (minimum 8 characters).
-        :param role: User role - "admin" or "user" (default: "user").
-        :param display_name: Optional display name.
-        :param avatar_url: Optional avatar URL.
-        :return: Created user object.
-        """
-        # Validation
-        if not username or len(username) < 3:
-            raise InvalidDataError("Username must be at least 3 characters")
-
-        if not password or len(password) < 8:
-            raise InvalidDataError("Password must be at least 8 characters")
-
-        # Validate role
-        try:
-            user_role = UserRole(role)
-        except ValueError as err:
-            raise InvalidDataError("Invalid role. Must be 'admin' or 'user'") from err
-
-        # Get built-in provider
-        builtin_provider = self.auth.login_providers.get("builtin")
-        if not builtin_provider or not isinstance(builtin_provider, BuiltinLoginProvider):
-            raise InvalidDataError("Built-in auth provider not available")
-
-        # Create user with password
-        user = await builtin_provider.create_user_with_password(username, password, role=user_role)
-
-        # Update optional fields if provided
-        if display_name or avatar_url:
-            updated_user = await self.auth.update_user(
-                user, display_name=display_name, avatar_url=avatar_url
-            )
-            if updated_user:
-                user = updated_user
-
-        self.logger.info("User created by admin: %s (role: %s)", username, role)
-        return user
-
-    @api_command("auth/user/enable", required_role="admin")
-    async def enable_user(self, user_id: str) -> dict[str, Any]:
-        """
-        Enable user account (admin only).
-
-        :param user_id: The user ID.
-        :return: Success status.
-        """
-        admin_user = get_current_user()
-        if not admin_user:
-            return {"success": False, "error": "Not authenticated"}
-
-        success = await self.auth.enable_user(user_id, admin_user)
-        return {"success": success}
-
-    @api_command("auth/user/disable", required_role="admin")
-    async def disable_user(self, user_id: str) -> dict[str, Any]:
-        """
-        Disable user account (admin only).
-
-        :param user_id: The user ID.
-        :return: Success status.
-        """
-        admin_user = get_current_user()
-        if not admin_user:
-            return {"success": False, "error": "Not authenticated"}
-
-        success = await self.auth.disable_user(user_id, admin_user)
-        return {"success": success}
-
-    @api_command("auth/user/delete", required_role="admin")
-    async def delete_user(self, user_id: str) -> dict[str, Any]:
-        """
-        Delete user account (admin only).
-
-        :param user_id: The user ID.
-        :return: Success status.
-        """
-        admin_user = get_current_user()
-        if not admin_user:
-            return {"success": False, "error": "Not authenticated"}
-
-        # Don't allow deleting yourself
-        if user_id == admin_user.user_id:
-            return {"success": False, "error": "Cannot delete your own account"}
-
-        # Delete user from database
-        try:
-            await self.auth.database.delete("users", {"user_id": user_id})
-            await self.auth.database.commit()
-            return {"success": True}
-        except Exception as e:
-            self.logger.exception("Error deleting user")
-            return {"success": False, "error": str(e)}
-
-    @api_command("auth/me")
-    async def get_current_user_info(self) -> User:
-        """Get current authenticated user information."""
-        current_user_obj = get_current_user()
-        if not current_user_obj:
-            raise AuthenticationRequired("Not authenticated")
-        return current_user_obj
-
-    @api_command("auth/user/update")
-    async def update_user_profile(
-        self,
-        user_id: str | None = None,
-        username: str | None = None,
-        display_name: str | None = None,
-        avatar_url: str | None = None,
-        password: str | None = None,
-        old_password: str | None = None,
-        role: str | None = None,
-        preferences: dict[str, Any] | None = None,
-    ) -> User:
-        """
-        Update user profile information.
-
-        Users can update their own profile. Admins can update any user including role and password.
-
-        :param user_id: User ID to update (optional, defaults to current user).
-        :param username: New username (optional).
-        :param display_name: New display name (optional).
-        :param avatar_url: New avatar URL (optional).
-        :param password: New password (optional, minimum 8 characters).
-        :param old_password: Current password (required when user updates own password).
-        :param role: New role - "admin" or "user" (optional, admin only).
-        :param preferences: User preferences dict (completely replaces existing, optional).
-        :return: Updated user object.
-        """
-        current_user_obj = get_current_user()
-        if not current_user_obj:
-            raise AuthenticationRequired("Not authenticated")
-
-        # Determine target user
-        if user_id and user_id != current_user_obj.user_id:
-            # Updating another user - requires admin
-            if current_user_obj.role != UserRole.ADMIN:
-                raise InsufficientPermissions("Admin access required")
-            target_user = await self.auth.get_user(user_id)
-            if not target_user:
-                raise InvalidDataError("User not found")
-            is_admin_update = True
-        else:
-            # Updating own profile
-            target_user = current_user_obj
-            is_admin_update = False
-
-        # Update role (admin only)
-        if role:
-            if not is_admin_update:
-                raise InsufficientPermissions("Only admins can update user roles")
-
-            try:
-                new_role = UserRole(role)
-            except ValueError as err:
-                raise InvalidDataError("Invalid role. Must be 'admin' or 'user'") from err
-
-            success = await self.auth.update_user_role(
-                target_user.user_id, new_role, current_user_obj
-            )
-            if not success:
-                raise InvalidDataError("Failed to update role")
-
-            # Refresh target user to get updated role
-            refreshed_user = await self.auth.get_user(target_user.user_id)
-            if not refreshed_user:
-                raise InvalidDataError("Failed to refresh user after role update")
-            target_user = refreshed_user
-
-        # Update basic profile fields
-        if username or display_name or avatar_url:
-            updated_user = await self.auth.update_user(
-                target_user,
-                username=username,
-                display_name=display_name,
-                avatar_url=avatar_url,
-            )
-            if not updated_user:
-                raise InvalidDataError("Failed to update user profile")
-            target_user = updated_user
-
-        # Update preferences if provided
-        if preferences is not None:
-            target_user = await self.auth.update_user_preferences(target_user, preferences)
-
-        # Update password if provided
-        if password:
-            if len(password) < 8:
-                raise InvalidDataError("Password must be at least 8 characters")
-
-            builtin_provider = self.auth.login_providers.get("builtin")
-            if not builtin_provider or not isinstance(builtin_provider, BuiltinLoginProvider):
-                raise InvalidDataError("Built-in auth not available")
-
-            if is_admin_update:
-                # Admin can reset password without old password
-                await builtin_provider.reset_password(target_user, password)
-                self.logger.info(
-                    "Password reset for user %s by admin %s",
-                    target_user.username,
-                    current_user_obj.username,
-                )
-            else:
-                # User updating own password - requires old password verification
-                if not old_password:
-                    raise InvalidDataError("old_password is required to change your own password")
-
-                # Verify old password and change to new one
-                success = await builtin_provider.change_password(
-                    target_user, old_password, password
-                )
-                if not success:
-                    raise AuthenticationFailed("Invalid current password")
-
-                self.logger.info("Password changed for user %s", target_user.username)
-
-        return target_user
-
-    @api_command("auth/tokens")
-    async def get_my_tokens(self, user_id: str | None = None) -> list[AuthToken]:
-        """
-        Get current user's auth tokens or another user's tokens (admin only).
-
-        :param user_id: Optional user ID to get tokens for (admin only).
-        :return: List of auth tokens.
-        """
-        current_user = get_current_user()
-        if not current_user:
-            return []
-
-        # If user_id is provided and different from current user, require admin
-        if user_id and user_id != current_user.user_id:
-            if current_user.role != UserRole.ADMIN:
-                return []
-            target_user = await self.auth.get_user(user_id)
-            if not target_user:
-                return []
-        else:
-            target_user = current_user
-
-        return await self.auth.get_user_tokens(target_user)
-
-    @api_command("auth/token/create")
-    async def create_user_token(self, name: str, user_id: str | None = None) -> dict[str, Any]:
-        """
-        Create a new long-lived access token for current user or another user (admin only).
-
-        Long-lived tokens are intended for external integrations and API access.
-        They expire after 10 years and do NOT auto-renew on use.
-
-        Short-lived tokens (for regular user sessions) are only created during login
-        and auto-renew on each use (sliding 30-day expiration window).
-
-        :param name: The name/description for the token (e.g., "Home Assistant", "Mobile App").
-        :param user_id: Optional user ID to create token for (admin only).
-        :return: Dictionary with the new token.
-        """
-        current_user = get_current_user()
-        if not current_user:
-            return {"success": False, "error": "Not authenticated"}
-
-        # If user_id is provided and different from current user, require admin
-        if user_id and user_id != current_user.user_id:
-            if current_user.role != UserRole.ADMIN:
-                return {"success": False, "error": "Admin access required"}
-            target_user = await self.auth.get_user(user_id)
-            if not target_user:
-                return {"success": False, "error": "User not found"}
-        else:
-            target_user = current_user
-
-        # Create a long-lived token (only long-lived tokens can be created via this command)
-        token = await self.auth.create_long_lived_token(target_user, name)
-        return {"success": True, "token": token}
-
-    @api_command("auth/token/revoke")
-    async def revoke_token(self, token_id: str) -> dict[str, Any]:
-        """
-        Revoke an auth token.
-
-        :param token_id: The token ID to revoke.
-        :return: Success status.
-        """
-        user = get_current_user()
-        if not user:
-            return {"success": False, "error": "Not authenticated"}
-
-        success = await self.auth.revoke_token(token_id, user)
-        return {"success": success}
-
-    @api_command("auth/logout")
-    async def logout(self) -> dict[str, Any]:
-        """
-        Logout current user by revoking the current token.
-
-        :return: Success status.
-        """
-        user = get_current_user()
-        if not user:
-            return {"success": False, "error": "Not authenticated"}
-
-        # Get current token from context
-        token = get_current_token()
-        if not token:
-            return {"success": False, "error": "No token in context"}
-
-        # Find and revoke the token
-        token_hash = hashlib.sha256(token.encode()).hexdigest()
-        token_row = await self.auth.database.get_row("auth_tokens", {"token_hash": token_hash})
-        if token_row:
-            await self.auth.revoke_token(token_row["token_id"], user)
-
-        return {"success": True}
-
-    @api_command("auth/user/providers")
-    async def get_my_providers(self) -> list[dict[str, Any]]:
-        """
-        Get current user's linked authentication providers.
-
-        :return: List of provider links.
-        """
-        user = get_current_user()
-        if not user:
-            return []
-
-        # Get provider links from database
-        rows = await self.auth.database.get_rows("user_auth_providers", {"user_id": user.user_id})
-        providers = [UserAuthProvider.from_dict(dict(row)) for row in rows]
-        return [p.to_dict() for p in providers]
-
-    @api_command("auth/user/unlink_provider", required_role="admin")
-    async def unlink_provider(self, user_id: str, provider_type: str) -> dict[str, Any]:
-        """
-        Unlink authentication provider from user (admin only).
-
-        :param user_id: The user ID.
-        :param provider_type: Provider type to unlink.
-        :return: Success status.
-        """
-        try:
-            await self.auth.database.delete(
-                "user_auth_providers", {"user_id": user_id, "provider_type": provider_type}
-            )
-            await self.auth.database.commit()
-            return {"success": True}
-        except Exception as e:
-            self.logger.exception("Error unlinking provider")
-            return {"success": False, "error": str(e)}
-
     async def _announce_to_homeassistant(self) -> None:
         """Announce Music Assistant Ingress server to Home Assistant via Supervisor API."""
         supervisor_token = os.environ["SUPERVISOR_TOKEN"]
@@ -1308,346 +903,3 @@ class WebserverController(CoreController):
                 )
         except Exception as err:
             self.logger.warning("Failed to announce to Home Assistant: %s", err)
-
-
-class WebsocketClientHandler:
-    """Handle an active websocket client connection."""
-
-    def __init__(self, webserver: WebserverController, request: web.Request) -> None:
-        """Initialize an active connection."""
-        self.webserver = webserver
-        self.mass = webserver.mass
-        self.request = request
-        self.wsock = web.WebSocketResponse(heartbeat=55)
-        self._to_write: asyncio.Queue[str | None] = asyncio.Queue(maxsize=MAX_PENDING_MSG)
-        self._handle_task: asyncio.Task[Any] | None = None
-        self._writer_task: asyncio.Task[None] | None = None
-        self._logger = webserver.logger
-        self._authenticated_user: Any = None  # Will be set after auth command or from Ingress
-        self._current_token: str | None = None  # Will be set after auth command
-        self._is_ingress = is_request_from_ingress(request)
-        self._events_unsub_callback: Any = None  # Will be set after authentication
-        # try to dynamically detect the base_url of a client if proxied or behind Ingress
-        self.base_url: str | None = None
-        if forward_host := request.headers.get("X-Forwarded-Host"):
-            ingress_path = request.headers.get("X-Ingress-Path", "")
-            forward_proto = request.headers.get("X-Forwarded-Proto", request.protocol)
-            self.base_url = f"{forward_proto}://{forward_host}{ingress_path}"
-
-    async def disconnect(self) -> None:
-        """Disconnect client."""
-        self._cancel()
-        if self._writer_task is not None:
-            await self._writer_task
-
-    async def handle_client(self) -> web.WebSocketResponse:
-        """Handle a websocket response."""
-        # ruff: noqa: PLR0915
-        request = self.request
-        wsock = self.wsock
-        try:
-            async with asyncio.timeout(10):
-                await wsock.prepare(request)
-        except TimeoutError:
-            self._logger.warning("Timeout preparing request from %s", request.remote)
-            return wsock
-
-        self._logger.log(VERBOSE_LOG_LEVEL, "Connection from %s", request.remote)
-        self._handle_task = asyncio.current_task()
-        self._writer_task = self.mass.create_task(self._writer())
-
-        # send server(version) info when client connects
-        server_info = self.mass.get_server_info()
-        await self._send_message(server_info)
-
-        # For Ingress connections, auto-create/link user and subscribe to events immediately
-        # For regular connections, events will be subscribed after successful authentication
-        if self._is_ingress:
-            await self._handle_ingress_auth()
-            self._subscribe_to_events()
-
-        disconnect_warn = None
-
-        try:
-            while not wsock.closed:
-                msg = await wsock.receive()
-
-                if msg.type in (WSMsgType.CLOSE, WSMsgType.CLOSING, WSMsgType.CLOSED):
-                    break
-
-                if msg.type != WSMsgType.TEXT:
-                    continue
-
-                self._logger.log(VERBOSE_LOG_LEVEL, "Received: %s", msg.data)
-
-                try:
-                    command_msg = CommandMessage.from_json(msg.data)
-                except ValueError:
-                    disconnect_warn = f"Received invalid JSON: {msg.data}"
-                    break
-
-                await self._handle_command(command_msg)
-
-        except asyncio.CancelledError:
-            self._logger.debug("Connection closed by client")
-
-        except Exception:
-            self._logger.exception("Unexpected error inside websocket API")
-
-        finally:
-            # Handle connection shutting down.
-            if self._events_unsub_callback:
-                self._events_unsub_callback()
-                self._logger.log(VERBOSE_LOG_LEVEL, "Unsubscribed from events")
-
-            try:
-                self._to_write.put_nowait(None)
-                # Make sure all error messages are written before closing
-                await self._writer_task
-                await wsock.close()
-            except asyncio.QueueFull:  # can be raised by put_nowait
-                self._writer_task.cancel()
-
-            finally:
-                if disconnect_warn is None:
-                    self._logger.log(VERBOSE_LOG_LEVEL, "Disconnected")
-                else:
-                    self._logger.warning("Disconnected: %s", disconnect_warn)
-
-        return wsock
-
-    async def _handle_command(self, msg: CommandMessage) -> None:
-        """Handle an incoming command from the client."""
-        self._logger.debug("Handling command %s", msg.command)
-
-        # Handle special "auth" command
-        if msg.command == "auth":
-            await self._handle_auth_command(msg)
-            return
-
-        # work out handler for the given path/command
-        handler = self.mass.command_handlers.get(msg.command)
-
-        if handler is None:
-            await self._send_message(
-                ErrorResultMessage(
-                    msg.message_id,
-                    InvalidCommand.error_code,
-                    f"Invalid command: {msg.command}",
-                )
-            )
-            self._logger.warning("Invalid command: %s", msg.command)
-            return
-
-        # Check authentication if required
-        if handler.authenticated or handler.required_role:
-            # For Ingress, user should already be set from _handle_ingress_auth
-            # For regular connections, user must be set via auth command
-            if self._authenticated_user is None:
-                await self._send_message(
-                    ErrorResultMessage(
-                        msg.message_id,
-                        AuthenticationRequired.error_code,
-                        "Authentication required. Please send auth command first.",
-                    )
-                )
-                return
-
-            # Set user and token in context for API methods
-            set_current_user(self._authenticated_user)
-            set_current_token(self._current_token)
-
-            # Check role if required
-            if handler.required_role == "admin":
-                if self._authenticated_user.role != UserRole.ADMIN:
-                    await self._send_message(
-                        ErrorResultMessage(
-                            msg.message_id,
-                            InsufficientPermissions.error_code,
-                            "Admin access required",
-                        )
-                    )
-                    return
-
-        # schedule task to handle the command
-        self.mass.create_task(self._run_handler(handler, msg))
-
-    async def _run_handler(self, handler: APICommandHandler, msg: CommandMessage) -> None:
-        try:
-            args = parse_arguments(handler.signature, handler.type_hints, msg.args)
-            result: Any = handler.target(**args)
-            if hasattr(result, "__anext__"):
-                # handle async generator (for really large listings)
-                items: list[Any] = []
-                async for item in result:
-                    items.append(item)
-                    if len(items) >= 500:
-                        await self._send_message(
-                            SuccessResultMessage(msg.message_id, items, partial=True)
-                        )
-                        items = []
-                result = items
-            elif asyncio.iscoroutine(result):
-                result = await result
-            await self._send_message(SuccessResultMessage(msg.message_id, result))
-        except Exception as err:
-            if self._logger.isEnabledFor(logging.DEBUG):
-                self._logger.exception("Error handling message: %s", msg)
-            else:
-                self._logger.error("Error handling message: %s: %s", msg.command, str(err))
-            err_msg = str(err) or err.__class__.__name__
-            await self._send_message(
-                ErrorResultMessage(msg.message_id, getattr(err, "error_code", 999), err_msg)
-            )
-
-    async def _writer(self) -> None:
-        """Write outgoing messages."""
-        # Exceptions if Socket disconnected or cancelled by connection handler
-        with suppress(RuntimeError, ConnectionResetError, *CANCELLATION_ERRORS):
-            while not self.wsock.closed:
-                if (process := await self._to_write.get()) is None:
-                    break
-
-                if callable(process):
-                    message: str = process()
-                else:
-                    message = process
-                self._logger.log(VERBOSE_LOG_LEVEL, "Writing: %s", message)
-                await self.wsock.send_str(message)
-
-    async def _send_message(self, message: MessageType) -> None:
-        """Send a message to the client (for large response messages).
-
-        Runs JSON serialization in executor to avoid blocking for large messages.
-        Closes connection if the client is not reading the messages.
-
-        Async friendly.
-        """
-        # Run JSON serialization in executor to avoid blocking for large messages
-        loop = asyncio.get_running_loop()
-        _message = await loop.run_in_executor(None, message.to_json)
-
-        try:
-            self._to_write.put_nowait(_message)
-        except asyncio.QueueFull:
-            self._logger.error("Client exceeded max pending messages: %s", MAX_PENDING_MSG)
-
-            self._cancel()
-
-    def _send_message_sync(self, message: MessageType) -> None:
-        """Send a message from a sync context (for small messages like events).
-
-        Serializes inline without executor overhead since events are typically small.
-        """
-        _message = message.to_json()
-
-        try:
-            self._to_write.put_nowait(_message)
-        except asyncio.QueueFull:
-            self._logger.error("Client exceeded max pending messages: %s", MAX_PENDING_MSG)
-
-            self._cancel()
-
-    async def _handle_auth_command(self, msg: CommandMessage) -> None:
-        """Handle WebSocket authentication command.
-
-        :param msg: The auth command message with access token.
-        """
-        # Extract token from args (support both 'token' and 'access_token' for backward compat)
-        token = msg.args.get("token") if msg.args else None
-        if not token:
-            token = msg.args.get("access_token") if msg.args else None
-        if not token:
-            await self._send_message(
-                ErrorResultMessage(
-                    msg.message_id,
-                    AuthenticationRequired.error_code,
-                    "token required in args",
-                )
-            )
-            return
-
-        # Authenticate with token
-        user = await self.webserver.auth.authenticate_with_token(token)
-        if not user:
-            await self._send_message(
-                ErrorResultMessage(
-                    msg.message_id,
-                    InvalidToken.error_code,
-                    "Invalid or expired token",
-                )
-            )
-            return
-
-        # Store authenticated user and token
-        self._authenticated_user = user
-        self._current_token = token
-        self._logger.info("WebSocket client authenticated as %s", user.username)
-
-        # Send success response
-        await self._send_message(
-            SuccessResultMessage(
-                msg.message_id,
-                {"authenticated": True, "user": user.to_dict()},
-            )
-        )
-
-        # Subscribe to events after successful authentication
-        self._subscribe_to_events()
-
-    async def _handle_ingress_auth(self) -> None:
-        """Handle authentication for Ingress connections (auto-create/link user)."""
-        ingress_user_id = self.request.headers.get("X-Remote-User-ID")
-        ingress_username = self.request.headers.get("X-Remote-User-Name")
-        ingress_display_name = self.request.headers.get("X-Remote-User-Display-Name")
-
-        if ingress_user_id and ingress_username:
-            # Try to find existing user linked to this HA user ID
-            user = await self.webserver.auth.get_user_by_provider_link(
-                AuthProviderType.HOME_ASSISTANT, ingress_user_id
-            )
-
-            if not user:
-                # Auto-create user for Ingress (they're already authenticated by HA)
-                user = await self.webserver.auth.create_user(
-                    username=ingress_username,
-                    role=UserRole.USER,
-                    display_name=ingress_display_name,
-                )
-                # Link to Home Assistant provider
-                await self.webserver.auth.link_user_to_provider(
-                    user, AuthProviderType.HOME_ASSISTANT, ingress_user_id
-                )
-
-            self._authenticated_user = user
-            self._logger.debug("Ingress user authenticated: %s", user.username)
-        else:
-            # No HA user headers - create/use system user for HA integration
-            # This allows the Home Assistant integration to connect via the internal network
-            # without needing Ingress headers (e.g., when using the Python client)
-            user = await self.webserver.auth.get_or_create_system_user(
-                username="homeassistant_system",
-                role=UserRole.USER,
-            )
-            self._authenticated_user = user
-            self._logger.debug("Ingress system user authenticated: %s", user.username)
-
-    def _subscribe_to_events(self) -> None:
-        """Subscribe to Mass events and forward them to the client."""
-        if self._events_unsub_callback is not None:
-            # Already subscribed
-            return
-
-        def handle_event(event: Any) -> None:
-            # event is MassEvent but we use Any to avoid runtime import
-            self._send_message_sync(event)
-
-        self._events_unsub_callback = self.mass.subscribe(handle_event)
-        self._logger.debug("Subscribed to events")
-
-    def _cancel(self) -> None:
-        """Cancel the connection."""
-        if self._handle_task is not None:
-            self._handle_task.cancel()
-        if self._writer_task is not None:
-            self._writer_task.cancel()
